BUGFIX: process lookup
We use /proc/<pid>/cmdline to lookup process names in run.py
and remote boot xinetd application bootComps.py

A process may have been gone between open(/proc/<PID>/cmdline)
and f.read().  In this case, we cannot read /proc/<PID>/cmdline
but this is not an error.  This commit fixes this bug.